Refactor ServerStatus component and enhance server health check logic
